Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    float s;
-    scanf("%f",&s);
-    printf("You entered: %.4f\n",s);
+    double s;
+    scanf("%lf",&s);
+    printf("You entered: %.4lf\n",s);
     return 0;
 }
